Implement download caching (experimental feature)
The idea is to have a list of download cache paths, where stored
filenames were the URL, and use only the last one on the list for
writing.

In this way, one can create a 'download snapshot' when Internet
access is available, then pack it, mount it somewhere read-only,
and reuse Stack somewhere else where the Internet is not available.

ToDo:
 * Take care of Git clones.
 * Think about a configuration where the entire Stackage snapshot
   is downloaded, so that future dependencies can use an existing
   'download snapshot'.