Implement the Stack System Root feature
(** work in progress **)

This change adds a new, optional configurable path that is available
via `STACK_SYSTEM_ROOT` environment, and also overridable via a
`stack-system-root` Stack config option, or a --stack-system-root
command line argument.

The Stack System Root path is an additional directory bearing
the same structure as `~/.stack`, but is considered a read-only
cache for anything that is stored under `~/.stack`. When Stack
runs, it first checks for availability of stuff under the Stack
System Root (downloads, installed GHC, indices, etc.), and only
afterward tries to either download, build, or install stuff to
`~/.stack`.